requirements.txt
matplotlib>=3.5.2
paho-mqtt>=1.6.1
requests>=2.27.1
